Adding test for checking the order
I don't know how Amazon does the sorting. So, I just tried to play around with getting the percentages of 5,4,3,2 and 1 star for each item and stored it into a list of string. I also got the number of people left the review for the item. Then I calculated the number of people who left 5 stars for each item and based on that I checked the ordering.